+++ audionimbus/src/effect/ambisonics/binaural.rs
@@ -77,6 +77,13 @@ impl AmbisonicsBinauralEffect {
         .into()
     }
 
+    /// Returns the number of tail samples remaining in an Ambisonics binaural effect’s internal buffers.
+    ///
+    /// Tail samples are audio samples that should be played even after the input to the effect has stopped playing and no further input samples are available.
+    pub fn tail_size(&self) -> usize {
+        unsafe { audionimbus_sys::iplAmbisonicsBinauralEffectGetTailSize(self.raw_ptr()) as usize }
+    }
+
     /// Resets the internal processing state of an ambisonics binaural effect.
     pub fn reset(&mut self) {
         unsafe { audionimbus_sys::iplAmbisonicsBinauralEffectReset(self.raw_ptr()) };