+++ audionimbus/src/effect/ambisonics/encode.rs
@@ -78,6 +78,13 @@ impl AmbisonicsEncodeEffect {
         .into()
     }
 
+    /// Returns the number of tail samples remaining in an Ambisonics encode effect’s internal buffers.
+    ///
+    /// Tail samples are audio samples that should be played even after the input to the effect has stopped playing and no further input samples are available.
+    pub fn tail_size(&self) -> usize {
+        unsafe { audionimbus_sys::iplAmbisonicsEncodeEffectGetTailSize(self.raw_ptr()) as usize }
+    }
+
     /// Resets the internal processing state of an ambisonics encode effect.
     pub fn reset(&mut self) {
         unsafe { audionimbus_sys::iplAmbisonicsEncodeEffectReset(self.raw_ptr()) };