+++ audionimbus/src/effect/ambisonics/rotation.rs
@@ -79,6 +79,13 @@ impl AmbisonicsRotationEffect {
         .into()
     }
 
+    /// Returns the number of tail samples remaining in an Ambisonics rotation effect’s internal buffers.
+    ///
+    /// Tail samples are audio samples that should be played even after the input to the effect has stopped playing and no further input samples are available.
+    pub fn tail_size(&self) -> usize {
+        unsafe { audionimbus_sys::iplAmbisonicsRotationEffectGetTailSize(self.raw_ptr()) as usize }
+    }
+
     /// Resets the internal processing state of an ambisonics rotation effect.
     pub fn reset(&mut self) {
         unsafe { audionimbus_sys::iplAmbisonicsRotationEffectReset(self.raw_ptr()) };