+++ audionimbus/src/effect/direct.rs
@@ -59,6 +59,13 @@ impl DirectEffect {
         .into()
     }
 
+    /// Returns the number of tail samples remaining in a direct effect’s internal buffers.
+    ///
+    /// Tail samples are audio samples that should be played even after the input to the effect has stopped playing and no further input samples are available.
+    pub fn tail_size(&self) -> usize {
+        unsafe { audionimbus_sys::iplDirectEffectGetTailSize(self.raw_ptr()) as usize }
+    }
+
     /// Resets the internal processing state of a direct effect.
     pub fn reset(&mut self) {
         unsafe { audionimbus_sys::iplDirectEffectReset(self.raw_ptr()) };