+++ audionimbus/src/effect/path.rs
@@ -65,6 +65,13 @@ impl PathEffect {
         .into()
     }
 
+    /// Returns the number of tail samples remaining in a path effect’s internal buffers.
+    ///
+    /// Tail samples are audio samples that should be played even after the input to the effect has stopped playing and no further input samples are available.
+    pub fn tail_size(&self) -> usize {
+        unsafe { audionimbus_sys::iplPathEffectGetTailSize(self.raw_ptr()) as usize }
+    }
+
     /// Resets the internal processing state of a path effect.
     pub fn reset(&mut self) {
         unsafe { audionimbus_sys::iplPathEffectReset(self.raw_ptr()) };