+++ audionimbus/src/effect/reflection.rs
@@ -113,6 +113,13 @@ impl ReflectionEffect {
         .into()
     }
 
+    /// Returns the number of tail samples remaining in a reflection effect’s internal buffers.
+    ///
+    /// Tail samples are audio samples that should be played even after the input to the effect has stopped playing and no further input samples are available.
+    pub fn tail_size(&self) -> usize {
+        unsafe { audionimbus_sys::iplReflectionEffectGetTailSize(self.raw_ptr()) as usize }
+    }
+
     /// Resets the internal processing state of a reflection effect.
     pub fn reset(&mut self) {
         unsafe { audionimbus_sys::iplReflectionEffectReset(self.raw_ptr()) };