+++ audionimbus/src/effect/virtual_surround.rs
@@ -69,6 +69,13 @@ impl VirtualSurroundEffect {
         .into()
     }
 
+    /// Returns the number of tail samples remaining in a virtual surround effect’s internal buffers.
+    ///
+    /// Tail samples are audio samples that should be played even after the input to the effect has stopped playing and no further input samples are available.
+    pub fn tail_size(&self) -> usize {
+        unsafe { audionimbus_sys::iplVirtualSurroundEffectGetTailSize(self.raw_ptr()) as usize }
+    }
+
     /// Resets the internal processing state of a virtual surround effect.
     pub fn reset(&mut self) {
         unsafe { audionimbus_sys::iplVirtualSurroundEffectReset(self.raw_ptr()) };